+++ tests/cli/MainCommand.test.ts
@@ -1,7 +1,7 @@
 import MainCommand from "../../src/cli/MainCommand";
 
 describe("MainCommand", () => {
-  it.skip("fails", () => {
-    expect(MainCommand("")).toBeFalsy();
-  });
+    it.skip("fails", () => {
+        expect(MainCommand("")).toBeFalsy();
+    });
 });