+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+-   Links to markdown files that are in the cosmere.json will now be rendered as Confluence links. (closes #12)
 -   Write cache even if upload is not successful. This saves time and helps with debugging.
 
 ## [0.15.0] - 2022-06-06